requirements.txt
allure-pytest>=2.15.0
allure_python_commons==2.15.0
fastapi==0.135.3
jsonschema==4.26.0
playwright==1.56.0
pydantic==2.12.5
pytest==8.4.2
python-dotenv==1.2.2
Requests==2.33.1
SQLAlchemy==2.0.49
